+++ packages/lumx-react/src/utils/react/unref.ts
@@ -1,7 +1,7 @@
 import { MaybeElementOrRef } from '@lumx/react/utils/type';
 
 /** Unref a react ref or element */
-export function unref(maybeElement: MaybeElementOrRef<HTMLElement>) {
-    if (maybeElement instanceof HTMLElement) return maybeElement;
+export function unref(maybeElement: MaybeElementOrRef<Element>) {
+    if (maybeElement instanceof Element) return maybeElement;
     return maybeElement?.current;
 }