+++ packages/lumx-react/src/utils/type/MaybeElementOrRef.ts
@@ -3,4 +3,4 @@ import type React from 'react';
 /**
  * Maybe a HTMLElement or a React ref of a HTMLElement
  */
-export type MaybeElementOrRef<E extends HTMLElement> = E | React.RefObject<E | null> | null | undefined;
+export type MaybeElementOrRef<E extends Element> = E | React.RefObject<E | null> | null | undefined;